fix: move MetricsTimer class to instrumentation package
